# tests/snapshots/openapi_v1.json
{
  "components": {
    "schemas": {
...
      "CareerPathItem": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "evidence_span_ids": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Span Ids",
            "type": "array"
          },
          "horizon": {
            "enum": [
              "short",
              "medium",
              "long"
            ],
            "title": "Horizon",
            "type": "string"
          }
        },
        "required": [
          "horizon",
          "action"
        ],
        "title": "CareerPathItem",
        "type": "object"
      },
      "CaseEvidenceResponse": {
        "additionalProperties": false,
        "properties": {
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "highest_stage": {
            "title": "Highest Stage",
            "type": "string"
          }
        },
        "required": [
          "case_id",
          "highest_stage"
        ],
        "title": "CaseEvidenceResponse",
        "type": "object"
      },
      "ConversationMessageResponse": {
        "additionalProperties": false,
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "kind": {
            "enum": [
              "intro",
              "brief",
              "progress",
              "checkpoint",
              "recovery",
              "result",
              "warning",
              "error"
            ],
            "title": "Kind",
            "type": "string"
          },
          "persona": {
            "enum": [
              "intent_consultant",
              "job_scout",
              "strategist",
              "pm"
            ],
            "title": "Persona",
            "type": "string"
          },
          "seq": {
            "minimum": 1.0,
            "title": "Seq",
            "type": "integer"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "seq",
          "persona",
          "display_name",
          "kind",
          "text",
          "stage"
        ],
        "title": "ConversationMessageResponse",
        "type": "object"
      },
      "EvidenceItem": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "evidence_span_id": {
            "title": "Evidence Span Id",
            "type": "string"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          }
        },
        "required": [
          "evidence_span_id",
          "content"
        ],
        "title": "EvidenceItem",
        "type": "object"
      },
      "ExecuteRunRequest": {
        "additionalProperties": false,
        "properties": {
          "plan_hash": {
            "maxLength": 64,
            "minLength": 64,
            "title": "Plan Hash",
            "type": "string"
          },
          "plan_version": {
            "minimum": 1.0,
            "title": "Plan Version",
            "type": "integer"
          }
        },
        "required": [
          "plan_version",
          "plan_hash"
        ],
        "title": "ExecuteRunRequest",
        "type": "object"
      },
      "FusionResponse": {
        "additionalProperties": false,
        "properties": {
          "implicit_max_weight": {
            "title": "Implicit Max Weight",
            "type": "number"
          }
        },
        "required": [
          "implicit_max_weight"
        ],
        "title": "FusionResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProductResult": {
        "additionalProperties": false,
        "properties": {
          "career_path": {
            "items": {
              "$ref": "#/components/schemas/CareerPathItem"
            },
            "title": "Career Path",
            "type": "array"
          },
          "recommended_roles": {
            "items": {
              "$ref": "#/components/schemas/RecommendationResult"
            },
            "title": "Recommended Roles",
            "type": "array"
          },
          "resume_strategy": {
            "items": {
              "$ref": "#/components/schemas/ResumeAdvice"
            },
            "title": "Resume Strategy",
            "type": "array"
          },
          "skill_gaps": {
            "items": {
              "$ref": "#/components/schemas/SkillGap"
            },
            "title": "Skill Gaps",
            "type": "array"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "summary"
        ],
        "title": "ProductResult",
        "type": "object"
      },
...
      "RankTraceResponse": {
        "additionalProperties": false,
        "properties": {
          "case_evidence": {
            "items": {
              "$ref": "#/components/schemas/CaseEvidenceResponse"
            },
            "title": "Case Evidence",
            "type": "array"
          },
          "case_ids": {
            "items": {
              "type": "string"
            },
            "title": "Case Ids",
            "type": "array"
          },
          "explicit_rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explicit Rank"
          },
          "explicit_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explicit Score"
          },
          "final_rank": {
            "title": "Final Rank",
            "type": "integer"
          },
          "implicit_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Implicit Confidence"
          },
          "implicit_rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Implicit Rank"
          },
          "implicit_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Implicit Score"
          },
          "implicit_weight": {
            "title": "Implicit Weight",
            "type": "number"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "final_rank",
          "implicit_weight"
        ],
        "title": "RankTraceResponse",
        "type": "object"
      },
...
      "RecommendationResult": {
        "additionalProperties": false,
        "properties": {
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "concise_explanation": {
            "title": "Concise Explanation",
            "type": "string"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceItem"
            },
            "title": "Evidence",
            "type": "array"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "listing_kind": {
            "default": "dataset_only",
            "enum": [
              "source_url",
              "dataset_only"
            ],
            "title": "Listing Kind",
            "type": "string"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "resume_evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceItem"
            },
            "title": "Resume Evidence",
            "type": "array"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "tier": {
            "enum": [
              "now_fit",
              "stretch_fit",
              "bridge_role"
            ],
            "title": "Tier",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "why_this_match": {
            "items": {
              "type": "string"
            },
            "title": "Why This Match",
            "type": "array"
          }
        },
        "required": [
          "job_id",
          "tier",
          "concise_explanation"
        ],
        "title": "RecommendationResult",
        "type": "object"
      },
      "RecoveryEventResponse": {
        "additionalProperties": false,
        "properties": {
          "attempt": {
            "title": "Attempt",
            "type": "integer"
          },
          "max_attempts": {
            "title": "Max Attempts",
            "type": "integer"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          }
        },
        "required": [
          "stage",
          "reason",
          "attempt",
          "max_attempts"
        ],
        "title": "RecoveryEventResponse",
        "type": "object"
      },
...
      "ResumeAdvice": {
        "additionalProperties": false,
        "properties": {
          "evidence_span_ids": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Span Ids",
            "type": "array"
          },
          "section": {
            "title": "Section",
            "type": "string"
          },
          "suggestion": {
            "title": "Suggestion",
            "type": "string"
          }
        },
        "required": [
          "section",
          "suggestion"
        ],
        "title": "ResumeAdvice",
        "type": "object"
      },
...
      "RunConversationResponse": {
        "additionalProperties": false,
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ConversationMessageResponse"
            },
            "title": "Messages",
            "type": "array"
          },
          "next_poll_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Poll Ms"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "status"
        ],
        "title": "RunConversationResponse",
        "type": "object"
      },
      "RunExplainResponse": {
        "additionalProperties": false,
        "properties": {
          "fusion": {
            "$ref": "#/components/schemas/FusionResponse"
          },
          "rank_trace": {
            "items": {
              "$ref": "#/components/schemas/RankTraceResponse"
            },
            "title": "Rank Trace",
            "type": "array"
          },
          "recovery_events": {
            "items": {
              "$ref": "#/components/schemas/RecoveryEventResponse"
            },
            "title": "Recovery Events",
            "type": "array"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "stage_durations_ms": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Stage Durations Ms",
            "type": "object"
          }
        },
        "required": [
          "run_id",
          "fusion"
        ],
        "title": "RunExplainResponse",
        "type": "object"
      },
      "RunResultResponse": {
        "additionalProperties": false,
        "properties": {
          "result": {
            "$ref": "#/components/schemas/ProductResult"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "status",
          "result"
        ],
        "title": "RunResultResponse",
        "type": "object"
      },
      "RunStatusResponse": {
        "additionalProperties": false,
        "properties": {
          "completed_stages": {
            "items": {
              "type": "string"
            },
            "title": "Completed Stages",
            "type": "array"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "execution_durability": {
            "title": "Execution Durability",
            "type": "string"
          },
          "plan_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Hash"
          },
          "plan_version": {
            "title": "Plan Version",
            "type": "integer"
          },
          "result_ready": {
            "title": "Result Ready",
            "type": "boolean"
          },
          "retry_after_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry After Ms"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "total_stages": {
            "default": 7,
            "title": "Total Stages",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "warning_codes": {
            "items": {
              "type": "string"
            },
            "title": "Warning Codes",
            "type": "array"
          }
        },
        "required": [
          "run_id",
          "session_id",
          "status",
          "result_ready",
          "execution_durability",
          "plan_version",
          "updated_at"
        ],
        "title": "RunStatusResponse",
        "type": "object"
      },
...
      "SkillGap": {
        "additionalProperties": false,
        "properties": {
          "evidence_span_ids": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Span Ids",
            "type": "array"
          },
          "gap": {
            "default": "",
            "title": "Gap",
            "type": "string"
          },
          "priority": {
            "default": "",
            "enum": [
              "low",
              "medium",
              "high",
              ""
            ],
            "title": "Priority",
            "type": "string"
          },
          "skill": {
            "title": "Skill",
            "type": "string"
          }
        },
        "required": [
          "skill"
        ],
        "title": "SkillGap",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Career-RAG",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/runs/{run_id}/conversation": {
      "get": {
        "operationId": "run_conversation_api_v1_runs__run_id__conversation_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunConversationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Conversation"
      }
    },
    "/api/v1/runs/{run_id}/execute": {
      "post": {
        "operationId": "execute_run_api_v1_runs__run_id__execute_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute Run"
      }
    },
    "/api/v1/runs/{run_id}/explain": {
      "get": {
        "operationId": "run_explain_api_v1_runs__run_id__explain_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunExplainResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Explain"
      }
    },
...
    "/api/v1/runs/{run_id}/result": {
      "get": {
        "operationId": "run_result_api_v1_runs__run_id__result_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Result"
      }
    },
    "/api/v1/runs/{run_id}/status": {
      "get": {
        "operationId": "run_status_api_v1_runs__run_id__status_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Status"
      }
    },
...
  }
}